test: move unit tests next to files
- Also reorg e2e test folder
