+++ e2e-tests/data/command-dictionary.xml
