+++ e2e-tests/fixtures/Dictionaries.ts
@@ -1,54 +1,56 @@
 import type { Locator, Page } from '@playwright/test';
 import { expect } from '@playwright/test';
+import { readFileSync } from 'fs';
 import { animals, colors, uniqueNamesGenerator } from 'unique-names-generator';
 
-export class Models {
+export class Dictionaries {
   readonly page: Page;
 
-  readonly jarPath: string;
-  readonly modelName: string;
-  readonly modelVersion: string;
+  readonly dictionaryBuffer: Buffer;
+  readonly dictionaryName: string;
+  readonly dictionaryPath: string;
 
-  readonly alertError: Locator;
   readonly confirmModal: Locator;
   readonly confirmModalDeleteButton: Locator;
   readonly createButton: Locator;
   readonly inputFile: Locator;
-  readonly inputName: Locator;
-  readonly inputVersion: Locator;
   readonly tableRow: Locator;
   readonly tableRowDeleteButton: Locator;
 
   constructor(page: Page) {
     this.page = page;
 
-    // TODO: Pull .jar file directly from built aerie project.
-    this.jarPath = 'tests/e2e/data/banananation-develop.jar';
-    this.modelName = uniqueNamesGenerator({ dictionaries: [colors, animals] });
-    this.modelVersion = '1.0.0';
+    this.dictionaryName = uniqueNamesGenerator({ dictionaries: [colors, animals] });
+    this.dictionaryPath = 'e2e-tests/data/command-dictionary.xml';
 
-    this.alertError = page.locator('.alert-error');
-    this.confirmModal = page.locator(`.modal:has-text("Delete Model")`);
-    this.confirmModalDeleteButton = page.locator(`.modal:has-text("Delete Model") >> button:has-text("Delete")`);
-    this.createButton = page.locator('text=Create');
+    const dictionaryFile = readFileSync(this.dictionaryPath);
+    const dictionaryXml = dictionaryFile.toString().replace(/GENERIC/, this.dictionaryName);
+    this.dictionaryBuffer = Buffer.from(dictionaryXml);
+
+    this.confirmModal = page.locator(`.modal:has-text("Delete Command Dictionary")`);
+    this.confirmModalDeleteButton = page.locator(
+      `.modal:has-text("Delete Command Dictionary") >> button:has-text("Delete")`,
+    );
+    this.createButton = page.locator(`button:has-text("Create")`);
     this.inputFile = page.locator('input[name="file"]');
-    this.inputName = page.locator('input[name="name"]');
-    this.inputVersion = page.locator('input[name="version"]');
-    this.tableRow = page.locator(`tr:has-text("${this.modelName}")`);
-    this.tableRowDeleteButton = page.locator(`tr:has-text("${this.modelName}") >> button[aria-label="Delete Model"]`);
+    this.tableRow = page.locator(`tr:has-text("${this.dictionaryName}")`);
+    this.tableRowDeleteButton = page.locator(
+      `tr:has-text("${this.dictionaryName}") >> button[aria-label="Delete Command Dictionary"]`,
+    );
   }
 
-  async createModel() {
+  async createDictionary() {
     await expect(this.tableRow).not.toBeVisible();
-    await this.fillInputName();
-    await this.fillInputVersion();
     await this.fillInputFile();
     await this.createButton.click();
     await this.tableRow.waitFor({ state: 'attached' });
     await expect(this.tableRow).toBeVisible();
   }
 
-  async deleteModel() {
+  /**
+   * @note Automatically cascade deletes any dependent expansion rules and expansion sets.
+   */
+  async deleteDictionary() {
     await expect(this.tableRow).toBeVisible();
     await expect(this.tableRowDeleteButton).not.toBeVisible();
 
@@ -69,24 +71,16 @@ export class Models {
 
   async fillInputFile() {
     await this.inputFile.focus();
-    await this.inputFile.setInputFiles(this.jarPath);
+    await this.inputFile.setInputFiles({
+      buffer: this.dictionaryBuffer,
+      mimeType: 'application/xml',
+      name: this.dictionaryName,
+    });
     await this.inputFile.evaluate(e => e.blur());
-  }
-
-  async fillInputName() {
-    await this.inputName.focus();
-    await this.inputName.fill(this.modelName);
-    await this.inputName.evaluate(e => e.blur());
-  }
-
-  async fillInputVersion() {
-    await this.inputVersion.focus();
-    await this.inputVersion.fill(this.modelVersion);
-    await this.inputVersion.evaluate(e => e.blur());
   }
 
   async goto() {
     await this.page.goto('/plans');
-    await this.page.goto('/models', { waitUntil: 'networkidle' });
+    await this.page.goto('/dictionaries', { waitUntil: 'networkidle' });
   }
 }
